test_requirements.txt
pytest
pytest-cov
pytest-asyncio
coveralls
python-telegram-bot